+++ docs/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Releases
 
-## v1.31.0
+## v1.31.1
 
 Released: ???
+
+Bugfixes:
+- Actualized loot pool items in the Fishing Profit Tracker as per Wiki https://wiki.hypixel.net/Fishing#Drops.
+- Added alert and party message for Treasure Dye catch (not tested for obvious reasons so might not work).
+- Fixed alert for Fiery Scuttler.
+- Fixed trigger message for Ice Essence catch.
+- Changed Alligator rarity to Legendary.
+
+## v1.31.0
+
+Released: 2025-03-18
 
 Features:
 - Adapted module to the new fishing update: